+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
-- Tracy to PSR bridge - custom levels are mapped to info instead of error if the logged value is not an exception
+- Tracy to PSR bridge
+	- custom levels are mapped to info instead of error if the logged value is not an exception
+	- custom levels are added to context key `custom_level`
 
 ## [1.3.1](https://github.com/orisai/nette-monolog/compare/1.3.0...1.3.1) - 2024-06-21